refactor: 기존 use-folder 훅 통합
