+++ src/pages/archive/components/MainFolderItem/MainFolderItem.tsx
@@ -1,11 +1,11 @@
 import Kebab from "@components/common/dropdown/Kebab";
 import { Text } from "@components/typography/Text";
-import { useFolderMark } from "@hooks/folder/use-folder";
 import { PATHS } from "@routes/paths";
 import { ArchiveNoteIcon, EmptyStarIcon } from "@svgs/index";
 import { getSafeColor } from "@utils/color";
 import { useState } from "react";
 import { useNavigate } from "react-router";
+import { useFolderMark } from "../../hooks/use-archive-folder";
 import { StarIcon } from "../StarIcon";
 import { DeleteFolderModal } from "../modal/DeleteFolderModal/DeleteFolderModal";
 import { EditFolderModal } from "../modal/EditFolderModal/EditFolderModal";
@@ -57,24 +57,18 @@ const MainFolderItem: React.FC<MainFolderItemProps> = ({
       {" "}
       <div
         className="flex flex-col relative w-[11.75rem] h-[11.75rem] p-6 pb-4 bg-white rounded-lg border border-gray-150 hover:bg-brand-20 cursor-pointer"
-        onClick={() => handleClick(id)}
+        onClick={() => handleClick(folderId)}
       >
         <div className="relative w-[3.75rem] h-[3.75rem] flex-shrink-0">
           <ArchiveMainFolderIcon fillColor={fillColor} />
 
-        <button
-          className="absolute top-8 right-[0.37rem] w-4 h-4 cursor-pointer"
-          onClick={handleToggleMark}
-        >
-          {markState ? <StarIcon /> : <EmptyStarIcon />}
-        </button>
-      </div>
+          <button className="absolute top-8 right-[0.37rem] w-4 h-4 cursor-pointer" onClick={handleToggleMark}>
+            {markState ? <StarIcon /> : <EmptyStarIcon />}
+          </button>
+        </div>
 
         {isArchive && (
-          <div
-            className="absolute top-6 right-4 cursor-pointer z-10"
-            onClick={(e) => e.stopPropagation()}
-          >
+          <div className="absolute top-6 right-4 cursor-pointer z-10" onClick={(e) => e.stopPropagation()}>
             <Kebab
               onSelect={(value) => {
                 if (value === "edit") setIsEditModalOpen(true);
@@ -84,14 +78,11 @@ const MainFolderItem: React.FC<MainFolderItemProps> = ({
           </div>
         )}
 
-      <div className="mt-4 w-[8.75rem] h-[2.25rem]">
-        <Text
-          variant="sub_heading2"
-          className="text-base-black text-ellipsis line-clamp-2 leading-tight"
-        >
-          {folderName}
-        </Text>
-      </div>
+        <div className="mt-4 w-[8.75rem] h-[2.25rem]">
+          <Text variant="sub_heading2" className="text-base-black text-ellipsis line-clamp-2 leading-tight">
+            {folderName}
+          </Text>
+        </div>
 
         <div className="flex items-center justify-between text-gray-400 mt-auto">
           <Text variant="sub_heading2" className="pr-8">
@@ -110,7 +101,7 @@ const MainFolderItem: React.FC<MainFolderItemProps> = ({
           onSubmit={handleEditFolder}
           folderName={folderName}
           folderColor={folderColor}
-          folderId={id}
+          folderId={folderId}
         />
       )}
       {isDeleteModalOpen && (
