+++ src/pages/archive/components/modal/EditFolderModal/EditFolderModal.tsx
@@ -1,11 +1,11 @@
 import { Text } from "@components/typography/Text";
-import { useColorUtils } from "@pages/archive/hooks/useColorUtils";
+import { usePatchFolders } from "@pages/archive/hooks/use-archive-folder";
+import { useColorUtils } from "@pages/archive/hooks/use-color-utils";
 import { colorMap } from "@styles/colorMap";
 import { ColorCircleCheckIcon, ColorCircleIcon } from "@svgs/index";
 import { isValidColor } from "@utils/color";
 import React, { useState, useRef } from "react";
 import { ArchiveFolderIcon } from "../../ArchiveFolderIcon";
-import { usePatchFolders } from "@pages/archive/hooks/use-folder";
 
 interface EditFolderModalProps {
   isOpen: boolean;
@@ -30,7 +30,7 @@ export const EditFolderModal: React.FC<EditFolderModalProps> = ({
   const [editedFolderName, setEditedFolderName] = useState(folderName);
 
   const [selectedColor, setSelectedColor] = useState<keyof typeof colorMap>(
-    isValidColor(folderColor) ? folderColor : "gray"
+    isValidColor(folderColor) ? folderColor : "gray",
   );
 
   const { mutate } = usePatchFolders();
@@ -95,10 +95,7 @@ export const EditFolderModal: React.FC<EditFolderModalProps> = ({
         </div>
 
         {/* Color Picker */}
-        <Text
-          variant="sub_heading3"
-          className="block text-gray-700 mb-[1.06rem]"
-        >
+        <Text variant="sub_heading3" className="block text-gray-700 mb-[1.06rem]">
           색상
         </Text>
         <div className="flex items-center mb-12 w-[21rem] h-[3.25rem]">
@@ -115,10 +112,7 @@ export const EditFolderModal: React.FC<EditFolderModalProps> = ({
                     style={{ fill: colorMap[color as keyof typeof colorMap] }}
                   />
                 ) : (
-                  <ColorCircleIcon
-                    className="w-5 h-5"
-                    style={{ fill: colorMap[color as keyof typeof colorMap] }}
-                  />
+                  <ColorCircleIcon className="w-5 h-5" style={{ fill: colorMap[color as keyof typeof colorMap] }} />
                 )}
               </button>
             ))}
@@ -135,11 +129,7 @@ export const EditFolderModal: React.FC<EditFolderModalProps> = ({
         </div>
 
         <div className="flex justify-end gap-2">
-          <button
-            type="button"
-            onClick={onClose}
-            className="h-8 px-5 py-1 rounded-md bg-gray-50 hover:bg-gray-100"
-          >
+          <button type="button" onClick={onClose} className="h-8 px-5 py-1 rounded-md bg-gray-50 hover:bg-gray-100">
             <Text variant="sub_heading2" className="text-gray-700">
               취소
             </Text>
